refactor module deployment template-operator constants into `tests/e2e`
